+++ python/GafferSceneUI/__init__.py
@@ -219,6 +219,7 @@ from . import CurvesInterpolationUI
 from . import MeshLightUI
 from . import AttributeProcessorUI
 from . import CurvesTangentsUI
+from . import CopyObjectUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.